clearState
