+++ src/features/embalm/stepContent/components/ReviewSarcophagus.tsx
@@ -3,7 +3,7 @@ import { useSarcoModal } from 'components/SarcoModal';
 import { ReviewSarcophagusTable } from './ReviewSarcophagusTable';
 import { SarcophagusSummaryFees } from './SarcophagusSummaryFees';
 import { SummaryErrorIcon } from './SummaryErrorIcon';
-import { CustomModal } from '../hooks/usePageBlock';
+import { PageBlockModal } from './PageBlockModal';
 
 export function ReviewSarcophagus() {
   const { SarcoModal, openModal, closeModal } = useSarcoModal();
@@ -14,7 +14,7 @@ export function ReviewSarcophagus() {
       spacing={6}
       mt={4}
     >
-      <CustomModal />
+      <PageBlockModal />
 
       <Text>
         Review your details below. Once you are ready, you can submit your transactions. Be aware,
